Created an Actions folder
Created an Actions folder inside src and filled it with four .js files:
authedUser, shared, tweets and users. tweets.js and users.js
currently contain one normal Action Creator each: tweets.js contains
(and exports) receiveTweets, which takes in a param of tweets and
returns an Action object with a type prop of RECEIVE_TWEETS, and also
a tweets prop. Similarly, users.js exports receiveUsers which returns
an Action object with a type prop of RECEIVE_USERS and a users prop.
shared.js exports an Action Creator that uses Thunk to handle the
asynchronous getInitialData request from the API. authedUser is empty.