Created setAuthedUser
Created the Action Creator setAuthedUser within authedUser.js .
From what I understand, it takes a user id as a param and returns an
Action object with a type prop and the user id as another prop. So
when it is used in the code, it should set whichever user id is
passed to it as the authenticated user, i.e. the user who is currently
using the app.